+++ src/utils/actionRenderers.ts
@@ -1,0 +1,5 @@
+import { GridProps } from "../types";
+
+export const selectActionRenderer = (actionName: string, gridProps: GridProps<any>) => {
+
+}